+++ src/design/plone/contenttypes/interfaces/servizio.py
@@ -334,7 +334,6 @@ class IServizio(model.Schema, IDesignPloneContentType):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "maximumSelectionSize": 10,
             "selectableTypes": ["UnitaOrganizzativa"],
             # "basePath": "/amministrazione/uffici",
         },