+++ src/design/plone/contenttypes/tests/test_ct_servizio.py
@@ -21,7 +21,6 @@ WIDGET_PROPERTY_CHECKS = {
         "selectableTypes": ["Pagina Argomento"],
     },
     "ufficio_responsabile": {
-        "maximumSelectionSize": 10,
         "selectableTypes": ["UnitaOrganizzativa"],
     },
     "area": {