+++ CHANGES.rst
@@ -7,7 +7,12 @@ Changelog
 
 - Remove selection limit in ufficio_responsabile field for Servizio.
   [foxtrot-dfm1]
-
+- Add new indexer "tassonomia_argomenti_uid" that indexes related Argomenti UIDs.
+  [cekk]
+- Change collection criteria to use new index.
+  [cekk]
+- Upgrade-step to convert old blocks with new criteria.
+  [cekk]
 
 5.0.3 (2022-12-07)
 ------------------